% Generated by roxygen2 (4.0.2): do not edit by hand
\name{factQuery}
\alias{factQuery}
\title{factQuery}
\usage{
factQuery(connexion, patient.list, restriction = "",
  aggregation = "patient")
}
\arguments{
\item{connexion}{: db connexion}

\item{patient.list}{: vector of patient_nums}

\item{restriction}{: string containing a SQL condition regarding the facts to return}

\item{aggregation}{: "patient" or "encounter"}
}
\value{
fact data.frame
}
\description{
Create a data.frame containing facts from a patient list in an i2b2 datacenter
}

