% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JDBCConnect}
\alias{JDBCConnect}
\title{JDBC Connect}
\usage{
JDBCConnect(config_path = "config_file")
}
\arguments{
\item{config_path}{Path to config_file (see getConfig() documentation)}
}
\value{
connection object
}
\description{
Create connection to a JDBC database
}

