% Generated by roxygen2 (4.0.2): do not edit by hand
\name{patientQuery}
\alias{patientQuery}
\title{patientQuery}
\usage{
patientQuery(connexion, query, aggregation = "patient")
}
\arguments{
\item{connexion}{: db connexion}

\item{query}{: string containing a selection criteria (e.g. "PATIENT_NUM = 1233405")}

\item{aggregation}{: "patient" or "encounter"}
}
\value{
patient data.frame
}
\description{
Create a data.frame containing patients characteristics selected by a specific query in an i2b2 datacenter
}

